backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.4
scikit-learn==1.5.1
google-generativeai==0.8.0
pydantic==2.9.0
